Move merkle tree spec file to proper place
